Fold braintrust adapter into importer package
