+++ plugins/packages/braintrust-importer/pyproject.toml
@@ -22,7 +22,11 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 keywords = ["ai-agents", "braintrust", "kitaru", "observability", "traces"]
-dependencies = ["kitaru>=0.22.0"]
+dependencies = [
+    "braintrust>=0.30,<1",
+    "httpx>=0.27,<1",
+    "kitaru>=0.23.0",
+]
 
 [project.urls]
 Homepage = "https://kitaru.ai"
