+++ plugins/packages/braintrust-importer/src/kitaru_braintrust_importer/__init__.py
@@ -1,1 +1,7 @@
 """Braintrust trace importer for Kitaru."""
+
+from .adapter import BraintrustAdapter
+
+__all__ = [
+    "BraintrustAdapter",
+]
