+++ tests/scripts/test_release_units.py
@@ -25,7 +25,6 @@ INVENTORY_PATH = REPO_ROOT / "release" / "release-units.toml"
 
 EXPECTED_UNITS = {
     "kitaru": "kitaru",
-    "braintrust": "kitaru-braintrust",
     "braintrust-importer": "kitaru-braintrust-importer",
     "evaluator": "kitaru-evaluator",
     "jsonl-importer": "kitaru-jsonl-importer",
@@ -120,7 +119,7 @@ def core_release_repo(
     return tmp_path, version
 
 
-def test_inventory_describes_core_and_fourteen_plugin_distributions() -> None:
+def test_inventory_describes_core_and_thirteen_plugin_distributions() -> None:
     inventory = load_inventory()
 
     assert {unit.slug: unit.distribution for unit in inventory.units} == EXPECTED_UNITS
@@ -644,7 +643,7 @@ def test_text_and_json_outputs_contain_the_same_unit_identities() -> None:
     assert all(unit.distribution in text_output for unit in inventory.units)
 
 
-def test_plugin_matrix_is_generated_from_the_fourteen_plugin_units() -> None:
+def test_plugin_matrix_is_generated_from_the_thirteen_plugin_units() -> None:
     matrix = build_plugin_matrix(load_inventory())
 
     assert matrix == {
@@ -834,7 +833,7 @@ def _run_cli(*arguments: str) -> subprocess.CompletedProcess[str]:
     [
         (["list"], "SLUG\tDISTRIBUTION\tVERSION\tDEFAULT\tTAG"),
         (["resolve", "--unit", "kitaru"], "python/kitaru/v"),
-        (["validate"], "Validated 15 release units."),
+        (["validate"], "Validated 14 release units."),
         (
             [
                 "propose-core-version",
